+++ apps/dsb-client-gateway-api/src/app/modules/dsb-client/dsb-client.interface.ts
@@ -147,7 +147,7 @@ export interface SendInternalMessageResponse {
   id: string;
 }
 
-export interface recipients {
+export interface Recipients {
   total: number;
   sent: number;
   failed: number;
@@ -166,7 +166,7 @@ export interface status {
 export interface SendMessageResponse {
   clientGatewayMessageId: string;
   did: string;
-  recipients: recipients;
+  recipients: Recipients;
   status: status[];
 }
